refactor: Refactor and export findProof when unlock erc20 to Ethereum.
