+++ .yarn/versions/c8026a30.yml
@@ -1,0 +1,3 @@
+releases:
+  "@near-eth/nep141-erc20": minor
+  rainbow-bridge-client-monorepo: minor
