+++ packages/nep141-erc20/src/bridged-nep141/index.js
@@ -4,3 +4,4 @@ export { default as getMetadata } from './getMetadata'
 export { default as deploy } from './deploy'
 export { initiate as sendToEthereum } from './sendToEthereum'
 export { recover } from './sendToEthereum'
+export { default as findProof } from './sendToEthereum/findProof'
